+++ mesonbuild/compilers/c.py
@@ -22,6 +22,7 @@ from .mixins.clike import CLikeCompiler
 from .mixins.ccrx import CcrxCompiler
 from .mixins.arm import ArmCompiler
 from .mixins.visualstudio import VisualStudioLikeCompiler
+from .mixins.gnu import GnuCompiler
 from .compilers import (
     gnu_winlibs,
     msvc_winlibs,
@@ -29,7 +30,6 @@ from .compilers import (
     ClangCompiler,
     Compiler,
     CompilerType,
-    GnuCompiler,
     ElbrusCompiler,
     IntelGnuLikeCompiler,
     IntelVisualStudioLikeCompiler,
