+++ mesonbuild/compilers/objcpp.py
@@ -18,7 +18,8 @@ import typing
 from ..mesonlib import EnvironmentException, MachineChoice
 
 from .mixins.clike import CLikeCompiler
-from .compilers import Compiler, ClangCompiler, GnuCompiler
+from .compilers import Compiler, ClangCompiler
+from .mixins.gnu import GnuCompiler
 
 class ObjCPPCompiler(CLikeCompiler, Compiler):
     def __init__(self, exelist, version, for_machine: MachineChoice, is_cross: bool, exe_wrap: typing.Optional[str]):
